# dashboard/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: MDDash Dashboard API
  version: 1.0.0
  description: API for experiment, simulation, compute, analysis, notebook, file, and publication workflows.
servers:
  - url: /
...
paths:
...
  /dash/api/experiments/{experiment_id}/tuner:
    parameters:
      - {$ref: '#/components/parameters/ExperimentId'}
    get:
      tags: [Tuner]
      summary: List tuner jobs
      operationId: listTunerJobs
      responses:
        '200':
          description: Tuner jobs
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/TunerJob'}}
        default: {$ref: '#/components/responses/Problem'}
    post:
      tags: [Tuner]
      summary: Start a tuner job
      operationId: startTunerJob
      parameters:
        - {name: simulation_path, in: query, schema: {type: string}}
        - {name: nsteps, in: query, schema: {type: integer, minimum: 1}}
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TunerJobRequest'}
      responses:
        '200':
          description: Existing tuner job
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TunerJob'}
        '201':
          description: Started tuner job
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TunerJob'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/experiments/{experiment_id}/tuner/{simulation_path}:
    parameters:
      - {$ref: '#/components/parameters/ExperimentId'}
      - {$ref: '#/components/parameters/SimulationPath'}
    get:
      tags: [Tuner]
      summary: Get a tuner job
      operationId: getTunerJob
      responses:
        '200':
          description: Tuner job
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TunerJob'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Problem'}
    delete:
      tags: [Tuner]
      summary: Delete a tuner job
      operationId: deleteTunerJob
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/experiments/{experiment_id}/tuner/{simulation_path}/stop:
    parameters:
      - {$ref: '#/components/parameters/ExperimentId'}
      - {$ref: '#/components/parameters/SimulationPath'}
    post:
      tags: [Tuner]
      summary: Stop a tuner job
      operationId: stopTunerJob
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/experiments/{experiment_id}/tuner/{simulation_path}/trials/{trial_id}/stdout:
    parameters:
      - {$ref: '#/components/parameters/ExperimentId'}
      - {$ref: '#/components/parameters/SimulationPath'}
      - {$ref: '#/components/parameters/TrialId'}
    get:
      tags: [Tuner]
      summary: Get tuner trial standard output
      operationId: getTunerTrialStdout
      responses:
        '200': {$ref: '#/components/responses/TextValue'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/experiments/{experiment_id}/tuner/{simulation_path}/trials/{trial_id}/stderr:
    parameters:
      - {$ref: '#/components/parameters/ExperimentId'}
      - {$ref: '#/components/parameters/SimulationPath'}
      - {$ref: '#/components/parameters/TrialId'}
    get:
      tags: [Tuner]
      summary: Get tuner trial standard error
      operationId: getTunerTrialStderr
      responses:
        '200': {$ref: '#/components/responses/TextValue'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  parameters:
    ExperimentId: {name: experiment_id, in: path, required: true, schema: {type: string, minLength: 1}}
    SimulationPath: {name: simulation_path, in: path, required: true, schema: {type: string, minLength: 1}}
...
    TrialId: {name: trial_id, in: path, required: true, schema: {type: string, minLength: 1}}
...
  responses:
...
    TextValue:
      description: Text represented as a JSON string
      content:
        application/json:
          schema: {type: string}
    NoContent: {description: Operation completed successfully}
...
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details. The HTTP status is carried by the response status line.
      required: [type, title, detail]
      properties:
        type: {type: string, format: uri, pattern: '^urn:mddash:'}
        title: {type: string, minLength: 1}
        detail: {type: string, minLength: 1}
        solution: {type: string, minLength: 1}
      additionalProperties: false
    Engine: {type: string, enum: [GMX, AMBER]}
    JobStatus: {type: string, enum: [UNKNOWN, PENDING, RUNNING, FINISHED, ERROR]}
...
    NullableString: {type: [string, 'null']}
...
    TunerTrial:
      type: object
      properties:
        id: {type: string}
        status: {$ref: '#/components/schemas/JobStatus'}
        performance: {type: [number, 'null']}
      additionalProperties: true
    TunerJob:
      type: object
      required: [id, experiment_id, simulation_path, created_at, nsteps, is_stopped, engine, tuner_status, trials, is_live]
      properties:
        id: {type: string}
        experiment_id: {type: string}
        simulation_path: {type: string}
        nsteps: {type: integer}
        error_message: {$ref: '#/components/schemas/NullableString'}
        created_at: {type: string, format: date-time}
        is_stopped: {type: boolean}
        engine: {$ref: '#/components/schemas/Engine'}
        tuner_status: {$ref: '#/components/schemas/JobStatus'}
        # True while the tuner job is not stopped and its status is non-terminal.
        is_live: {type: boolean}
        sim_length_ns: {type: [number, 'null']}
        trials: {type: array, items: {$ref: '#/components/schemas/TunerTrial'}}
      additionalProperties: false
    TunerJobRequest:
      type: object
      required: [simulation_path, nsteps]
      properties:
        simulation_path: {type: string, minLength: 1}
        nsteps: {type: integer, minimum: 1}
      additionalProperties: false
...